backend2/requirements.txt
Flask
requests
Flask-Cors
Flask-SQLAlchemy
Werkzeug
flask-jwt-extended
google-cloud-storage
psycopg2-binary

